libiconv: Add NMake build files for libiconv-1.17
This adds a set of NMake Makefiles for Visual Studio 2008~2022 for x86, x64 and
ARM64 builds of libiconv as DLLs, with pre-configured configuration headers as
well as gnulib headers that are used in the build, as well as patches required
for pre-2013/2015 Visual Studio.

This supports building libiconv and libcharset as DLLs only.