libffi: Add NMake Makefile set for 3.4.4
This adds a set of NMake Makefiles to build libffi-3.4.4 as a DLL in x86, x64
and ARM64 flavors as a DLL, for at least Visual Studio 2015 and later.

Earlier Visual Studio versions may rqeuire updating the code to be
C89-compliant, which will be done in a subsequent update.