gettext: Add prelminary set of NMake build files for 0.21.1
This adds a set of NMake Makefiles and preconfigured headers needed to build
the gettext libraries (libintl, libasprintf, libtextstyle, libgettextpo) as
DLLs as well the tools that come with gettext-runtime and gettext-tools.

From 0.21.1, only Visual Studio 2015 and later are supported (x86, x64 and
ARM64 builds).

This includes a preliminary 'install' target that will copy the headers, built
tools and libraries (DLL and import .lib's) in a location specified by
$(PREFIX), or in a default location if $(PREFIX) is unspecified.

The data files and auxiliary files are not yet installed at this point.