gettext/0.22: Add set of NMake Makefiles for gettext-runtime
This adds a set of NMake Makefiles that are used to build the items that
reside in gettext-runtime, including libintl and libasprintf as DLLs, as well
as the C# GNU.gettext.dll and the tools in gettext-runtime/src.

Note that although there are settings for libtextstyle and for the items in
gettext-tools, these are not yet ready as those are currently carried over
from the NMake Makefiles in 0.21.1, and are not made ready for the 0.22
release yet.