gettext/0.22: Allow specifying libiconv directories
One may wish to specify directories where the libiconv
installation/binaries reside, or even specify paths in which its
headers, libraries and DLLs can be found, so this adds options to the
NMake Makefiles to specify and honor these items, in addition to looking
in the specified $(PREFIX), if applicable.

Also fine-tune some of the include orders as a result of this, and fix
up some leftbehinds.