gettext/0.22.4: Add pre-configured headers for gettext-tools
This includes the various pre-configured header files for building the
items in gettext-tools, in particular the GNULib parts that are built in
the process, for x86, x64 and ARM64 builds under Visual Studio 2015 and
later.

This will make the gettext-0.22.4.sources ready to build with NMake, as
the NMake Makefiles doesn't seem necessary to be updated at this point.